create user and activate account using email

+++ src/services/auth.service.ts
@@ -70,6 +70,48 @@ class AuthService {
     return admin;
   }
 
+  public async signUp(dto: Partial<IUser>): Promise<IUser> {
+    const userFromDb = await userRepository.getOneByParams({
+      email: dto.email,
+    });
+    if (userFromDb) {
+      throw new ApiError("User already exists", 400);
+    }
+
+    const hashedPassword = await passwordService.hash(
+      dto.password,
+      EUserRole.SELLER,
+    );
+    const role = await roleRepository.getOneByParams({
+      role: EUserRole.SELLER,
+    });
+
+    const user = await userRepository.create({
+      ...dto,
+      password: hashedPassword,
+      _roleId: role._id,
+    });
+
+    const actionToken = actionTokenService.createActionToken(
+      { userId: user._id, roleId: role._id },
+      EActionTokenType.ACTIVATE_ACCOUNT,
+    );
+
+    await Promise.all([
+      actionTokenRepository.createActionToken({
+        actionToken,
+        _userId: user._id,
+        tokenType: EActionTokenType.ACTIVATE_ACCOUNT,
+      }),
+      emailService.sendMail(dto.email, EEmailAction.ACTIVATE_ACCOUNT, {
+        userName: dto.userName,
+        actionToken,
+      }),
+    ]);
+
+    return user;
+  }
+
   public async signUpVerifyAdmin(actionToken: string) {
     const payload = actionTokenService.checkActionToken(
       actionToken,
@@ -92,6 +134,33 @@ class AuthService {
       userRepository.updateById(admin._id, {
         accountStatus: EAccountStatus.ACTIVE,
         accountType: EAccountType.PREMIUM,
+      }),
+
+      actionTokenRepository.deleteActionTokenByParams({ actionToken }),
+    ]);
+  }
+
+  public async signUpVerify(actionToken: string) {
+    const payload = actionTokenService.checkActionToken(
+      actionToken,
+      EActionTokenType.ACTIVATE_ACCOUNT,
+    );
+    if (!payload) {
+      throw new ApiError("Not valid token", 400);
+    }
+
+    const entity = await actionTokenRepository.getActionTokenByParams({
+      actionToken,
+    });
+    if (!entity) {
+      throw new ApiError("Not valid token", 400);
+    }
+
+    const user = await userRepository.getOneByParams({ _id: entity._userId });
+
+    await Promise.all([
+      userRepository.updateById(user._id, {
+        accountStatus: EAccountStatus.ACTIVE,
       }),
 
       actionTokenRepository.deleteActionTokenByParams({ actionToken }),
